+++ NEWS.md
@@ -1,9 +1,13 @@
-# dataprocr2 0.1.4
+# dataproc.iquizoo 0.2.0
+
+* Rename package from "dataprocr2" to "dataproc.iquizoo". This is recommended because all the work is related to www.iquizoo.com.
+
+# dataproc.iquizoo 0.1.4
 
 * Added a utility function `calc_sdt()` to calculate signal detection theory related indices.
 * Fixed an issue related to perfect accuracy in `drm()`.
 
-# dataprocr2 0.1.3
+# dataproc.iquizoo 0.1.3
 
 * Added a `NEWS.md` file to track changes to the package.
 * Added `pc_all` index, meaning overal percent of correct, to `wxpred` function.